requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-allauth==0.63.6
django-cors-headers==4.6.0
django-environ==0.11.2
django-otp==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
multidict==6.0.5
packaging==24.1
paystackease==2.3.0
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
yarl==1.9.4
